eventbot: XMPP event stream bot
eventbot is a chat bot for command line style interaction with streams
of events in a group chat. Examples might be querying and selecting
events from riemann or booting a minimal linux container and grepping
through stdout.

eventbot is built around flow control: events are throttle and dropped
to avoid overwhelming systems, networks and humans.